Renamed some files to prevents some warnings with msvc
