+++ source/Core/Castor3D/Text/TextScene.cpp
@@ -12,6 +12,7 @@
 #include "Castor3D/Render/RenderLoop.hpp"
 #include "Castor3D/Text/TextAnimatedObjectGroup.hpp"
 #include "Castor3D/Text/TextBackground.hpp"
+#include "Castor3D/Text/TextBaseOverlay.hpp"
 #include "Castor3D/Text/TextCamera.hpp"
 #include "Castor3D/Text/TextFog.hpp"
 #include "Castor3D/Text/TextGeometry.hpp"
@@ -19,7 +20,6 @@
 #include "Castor3D/Text/TextMaterial.hpp"
 #include "Castor3D/Text/TextMesh.hpp"
 #include "Castor3D/Text/TextParticleSystem.hpp"
-#include "Castor3D/Text/TextOverlay.hpp"
 #include "Castor3D/Text/TextRenderTarget.hpp"
 #include "Castor3D/Text/TextRenderWindow.hpp"
 #include "Castor3D/Text/TextSampler.hpp"
