+++ source/Core/CastorUtils/Platform/Android/AndroidAligned.cpp
