+++ source/Core/CastorUtils/Platform/Android/AndroidCpuInformations.cpp
