+++ source/Core/CastorUtils/Platform/Android/AndroidDebug.cpp
