+++ source/Core/CastorUtils/Platform/Android/AndroidDynamicLibrary.cpp
