+++ source/Core/CastorUtils/Platform/Android/AndroidFile.cpp
