+++ source/Core/CastorUtils/Platform/Android/AndroidLoggerConsole.cpp
