+++ source/Core/CastorUtils/Platform/Android/AndroidUtils.cpp
