+++ source/Core/CastorUtils/Platform/Linux/LinuxAligned.cpp
