+++ source/Core/CastorUtils/Platform/Linux/LinuxCpuInformations.cpp
