+++ source/Core/CastorUtils/Platform/Linux/LinuxDebug.cpp
