+++ source/Core/CastorUtils/Platform/Linux/LinuxDynamicLibrary.cpp
