+++ source/Core/CastorUtils/Platform/Linux/LinuxFile.cpp
