+++ source/Core/CastorUtils/Platform/Linux/LinuxLoggerConsole.cpp
