+++ source/Core/CastorUtils/Platform/Linux/LinuxUtils.cpp
