+++ source/Core/CastorUtils/Platform/MacOS/MacOSAligned.cpp
