+++ source/Core/CastorUtils/Platform/MacOS/MacOSDebug.cpp
