+++ source/Core/CastorUtils/Platform/MacOS/MacOSFile.cpp
