+++ source/Core/CastorUtils/Platform/MacOS/MacOSLoggerConsole.cpp
