+++ source/Core/CastorUtils/Platform/MacOS/MacOSUtils.cpp
