+++ source/Core/CastorUtils/Platform/Win32/Win32Aligned.cpp
